Rechunk rfftfreq/fftfreq if needed (#2473)
* Store normalized chunks without adjustment

Normalize the chunks before handling the excess chunk. This way we know
that `chunks` should still match what the user intended. Thus we can
reuse this in later steps where the `chunks` will be needed.

* Rechunk fftfreq and rfftfreq

To ensure the returned Dask Array has the chunks specified, make sure to
rechunk after restructuring the Dask Array should it be needed. If the
chunks are already correct, simply skip this step.

* Test that rfftfreq/fftfreq have the right chunks