Rewrite fftfreq/rfftfreq to use map_blocks
In order to clean up and simplify the `fftfreq` and `rfftfreq` functions
a bit, split them up and rewrite them in terms of `map_blocks` on a
Python function to create the resultant Dask Array. This removes many of
the odd distinctions that occurred by trying to combine the real and
complex cases. Also it avoids the need to perform weird chunk
manipulation either at array creation or subsequently (including the
call to `rechunk`). Not only are these functions simpler and easier to
reason about, but it appears they are a bit faster than the original
implementations.