Flatten rfftfreq function
There is really no need for `map_blocks` in `rfftfreq` as the function
literally consists of dividing by a scalar. So this is more overhead and
code complexity than just doing the division directly with Dask. Hence
we drop the closure and `map_blocks` from `rfftfreq` to simplify things
a bit.