XFail test_interrupt with pytest-xdist (#2483)
This test started failing periodically again, but only when used with
`pytest-xdist`. I can't reproduce it locally, and I'm not sure that
`pytest-xdist` is actually the culprit. For now we xfail for xdist runs.
This test has used up too much dev time, and is still flakier than
baklava. Frustrating.