Add `subok` kwarg to `da.from_array`
Mirrors the `subok` kwarg in `np.array`. If True, array-like's will be
passed through as chunks unchanged. Default is False. This allows
avoiding the `np.asarray` call, making it easier to work with non numpy
arrays as chunks.

To make this work we change `getarray` to a more generic `getter`
function and add a `subok` parameter. Alternatively we could have kept
the same format and created more getter functions, but this would have
led to an explosion of functions for all the different combinations
(which seemed less maintainable).