Add fire_and_forget function (#1221)
* Add fire_and_forget function

Run tasks at least once, even if we release the futures

Under normal operation Dask will not run any tasks for which there is not an
active future (this avoids unnecessary work in many situations).  However
sometimes you want to just fire off a task, not track its future, and expect it
to finish eventually.  You can use this function on a future or collection of
futures to ask Dask to complete the task even if no active client is tracking
it.

The results will not be kept in memory after the task completes (unless there
is an active future) so this is only useful for tasks that depend on side
effects.

Parameters
----------

obj: Future, list, dict, dask collection
    The futures that you want to run at least once

Examples
--------

    >>> fire_and_forget(client.submit(func, *args))

Fixes https://github.com/dask/distributed/issues/1205

* Make fire and forget tests more reliable on travis-ci