Add an implementation of nonzero
Provides a basic implementation of NumPy's nonzero for Dask. Makes use
of `compress` to make it work. As `compress` seems to be eagerly
evaluated, this means `nonzero` is as well. So it is worth keeping that
in mind. However it is possible the `compress` implementation could be
revisited and made lazy, which would ensure `nonzero` is lazy as well.