Improve `Client.get_versions` (#1279)
* get_versions works with different cloudpickles

Previously we this required serializing a function with a closure, which
would error if differing versions of cloudpickle existed across the
cluster. We now just call `distributed.versions.get_versions` directly.

* Improve error message for get_versions

Previously this would error eagerly informing about only one mismatch
when there may be many. This led to playing whack-a-mole with version
compatibility. We now throw a nice error message enumerating all the
mismatched packages.