Handle axis as None in compress
According to NumPy's documentation, when `compress` gets `axis` as
`None`, it simply flattens the array and applies the same operation on
it. Given this, it appears to be a simple matter to transform this case
into an already supported case for `compress`; namely, a flattened array
where `axis` is `0` instead.