In compress, use Dask to expand condition array (#2545)
To prep for making `compress` a lazy function, handle the `condition`
array argument as a Dask Array as long as possible. Unfortunately it
still needs to be converted to a NumPy Array to create the needed slice
to apply to the input array `a`. However this little rewrite should make
it easier to focus on just getting that portion of code into Dask Arrays
instead of NumPy Arrays.