Support dask array conditional in compress (#2555)
* Support dask array conditional in compress

Compress now works with dask arrays as conditionals without calling
compute on them. If the conditional is concrete we take an optimized
path and return an array of known shape. If it's lazy we return an
array of unknown shape along the specified axis.

* Add extract

Implements `da.extract` using `da.compress`.