Stacking Dask Arrays with unknown dim doesn't work
Appears that Dask Array's `stack` does not work with Dask Arrays that
have an unknown dimension length. While that makes sense in the general
case (we don't know their lengths), it doesn't apply to this case where
we do know their lengths. In any event, we can update our tests
accordingly as we do here.