Rewrite nonzero to use argwhere
As `argwhere` basically does all the work up to the point of getting
transposed and returning `tuple`s, it does much of the work that
`nonzero` was doing. Given this, it makes sense to rewrite `nonzero` in
terms of `argwhere`. Since `argwhere` simply contains one array with all
of the data, it may be more useful in some circumstances than `nonzero`
for use with Dask Arrays.