In argwhere, improve handling of trivial arrays
If any array had trivial dimensions, an issue would occur in the
`reshape` step, which blocked getting a result for `argwhere`. To fix
this, we add a workaround that provides the same behavior by applying
`ravel` to each chunk. The result is an implementation that works just
as well on trivial and non-trivial arrays.