Fix count_nonzero's test scalar comparison
As Dask Array represents `count_nonzero` as a zero dimensional array
when all axes are collapsed before it is computed while NumPy represents
it as a Python integer, simply compare the results between the two in
this case instead of comparing the arrays. This will ensure the values
are correct even if they may represent the result slightly differently.