Refactor out a nonzero function
Creates an internal nonzero function that can be used to work on numeric
arrays, string data, and generic object arrays. The latter two cases may
be notably less efficient due to the use of NumPy's `vectorize`.
Unfortunately that seems to be the best way to solve those issues, so
that is what we go with for now. Can always be replaced by a more
efficient method if one is found.