Test argwhere and count_nonzero on strs and objs
As these are using our nonzero function directly, these are the two
function we want to be sure work for the different type cases. Hence the
additional tests on them. Higher level functions based off of these will
do the right thing as a consequence of these lower level functions'
behavior.