Drop 0-d array case from count_nonzero axis tests
The axis test for `count_nonzero` is suppose to exercise a variety of
non-trivial axes. This does not mesh with 0-d array case as it cannot
have any axes specified. Hence we drop this case from the
`count_nonzero` axis test. After all it is still tested when the axes
are trivial.