Generalize isnonzero vectorization handling
Try a simple test conversion for the type in question to see if NumPy
can in fact convert it to `bool`. From quick testing, this seems to
properly pass through all numeric types and object arrays. It catches
string types correctly and passes them on to the vectorization function.
So it seems to retain the behavior had before, but a better forward
looking manner.

The nice thing about this check is it actually is testing the operation
Dask will need to perform to make the conversion happen. So it is a good
indicator if it can proceed or not. Further if NumPy fixes any cases
that currently fails (or vice versa), we should still be able to catch
them correctly and handle them appropriately. All of this being done
before generating the Dask array.