Clear `_cached_keys` on name change in dask.array
Previously a `_cached_keys` attribute was added to `dask.array` to store
the result of the `_keys` method for fast recompute. Later, mutating the
`name`/`dask` attributes of a `dask.array` was supported. Unfortunately,
mutating these attributes wouldn't clear the cache, which would lead to
hard to find bugs.

To fix this, we do the following:

- Make `Base` define `__slots__`, so that the `__slots__` attribute of
  `Array` is respected. Before the `__slots__` on `Array` were ignored,
  making `_cached_keys` an unexpected source of state (wasn't declared,
  and the declaration was thought to be respected)

- Make `name` a property, and clear the cache when set

- Add a test for the cache behavior.