Include computed collection within sharedict in delayed
When delayed is given a collection it optimizes and finalizes the
collection's graph before adding it into its own graph.  Because delayed
uses a sharedict it gives this new subgraph a name.  Previously we did
not provide a good name for graphs in this case and instead relied on
the id number of the dict.  Now we watch for this case more carefully.