Fixes for elemwise dtype inference for 0d arrays
This fixed #2574 and #2575. When given a 0d dask array, it now
passes a 1d dummy numpy array to `apply_infer_dtype`. This assumes that
the operation is capable of handling broadcasting - in particular if
elemwise is given a numpy 0d array and a dask 0d array, it will use
mixed dimensionalities in dtype inference. I had a quick look through
the array code and didn't see any problematic cases, and the unit tests
pass.

Additionally, when a 0d dask array is used, the result of
apply_infer_dtype might not match the actual returned type if numpy
type rules are in use. In this case, the operation is wrapped to cast
the output to the inferred type, to ensure that the Array has a dtype
that matches the end result.

I added a unit test to cover #2575. I deleted some existing code from
the test_elemwise_on_scalars test that seemed to have nothing to do with
elemwise.