Add ediff1d (#2609)
* Add ediff1d

Provides a Dask Array implementation of NumPy's `ediff1d`, which
bootstraps itself off of `diff`.

* Test ediff1d

Try a variety of parameters with `ediff1d` and compare the results between
the Dask Array implementation and the NumPy implementation.

* Include ediff1d in the Dask Array API docs

* Rewrite ediff1d without diff

Since there is no need to consider taking multiple differences or
working on a particular axis, `ediff1d` can be written more simply
without needing to use `diff`. This change makes use of this fact.