In norm handle keepdims for svd-based results
There were a few cases in `norm` where `svd` was used and we forgot to
add back in the dimensions we removed when `keepdims` was `True`.
However this case is a little trickier as we want to operate on a vector
of singular values. However we need to get back two singleton
dimensions. In order to manage this, we take the result from `svd` and
tack on a singleton dimension before passing it to the appropriate
reduction function. That way the reduction function will either keep 2
singleton dimensions (the one we added and the one from the vector) or
it will get rid of both of them. No additional branch is required in
this case.