Skip some apply_along_axis tests on old NumPys
Before NumPy 1.13.0, it was not permissible for `apply_along_axis` to
take a function that returned an array with dimensionality greater than
1-D. As such we should not bother testing cases where `apply_along_axis`
is given a function that returns something with higher dimensionality if
we don't have a NumPy version that can handle such a result and be
compared to. Thus we skip this case for any NumPy prior to 1.13.0.