Add apply_along_axis (#2698)
* Add apply_along_axis to Dask Array

Implements a Dask Array equivalent of `apply_along_axis`. Does this by
leveraging a NumPy object Array to store all results before stacking
them all back together as one unified Dask Array.

* Export apply_along_axis to Dask Array's public API

* Test Dask Array's apply_along_axis

Provide some tests for the Dask Array implementation of
`apply_along_axis`, which compares it to the NumPy implementation. Try a
range of different functions and axes they are applied along. Also
explores functions that return something with some dimensionality to
them.

* Document Dask Array's apply_along_axis

* Rewrite apply_along_axis to use map_blocks

Previously the `apply_along_axis` implementation went through the Dask,
pulling out each 1D array, and applying the user provided function to
them. However this could create a rather large graph with a significant
amount of overhead for large Dask Arrays. A better solution ends up
being rechunking the data into 1D array chunks the size of which the
user provided function will operate on. Then simply using `map_blocks`
to apply the user provided function to each chunk. To ensure the right
size is determined for the result, some test data is run through the
user provided function first to estimate the results shape to provide to
`map_blocks`.

* Skip some apply_along_axis tests on old NumPys

Before NumPy 1.13.0, it was not permissible for `apply_along_axis` to
take a function that returned an array with dimensionality greater than
1-D. As such we should not bother testing cases where `apply_along_axis`
is given a function that returns something with higher dimensionality if
we don't have a NumPy version that can handle such a result and be
compared to. Thus we skip this case for any NumPy prior to 1.13.0.

* Use NumPy's apply_along_axis inside ours

To simplify our implementation of `apply_along_axis` and particularly
our wrapper function, just call `apply_along_axis` within our wrapper
function. This should allow it to be a bit more flexible, standardized,
and hopefully more efficient.

* Rechunk axis only in apply_along_axis

Instead of rechunking all other axes not being computed along to 1,
simply preserve their existing chunking. Only focus on having one chunk
for the axis that will be operated on by `apply_along_axis`. Due to
recent improvements under the hood in our wrapper function, this should
work seamlessly despite the fact there are some non-trivial dimensions.

* Raise for unsupported result shape

In the event a user has a NumPy older than 1.13.0, we cannot support
multiple dimension from the `func1d` function given to
`apply_along_axis`. So in this case, raise a `ValueError` just as old
version of NumPy would. Update the tests to make sure this error is
thrown.