Use apply_along_axis name in Dask
Previously an internal function was being used to name the
`apply_along_axis` operation in the Dask graph. This is a bit confusing
and not really what we want the user to see. So instead use
`apply_along_axis` as the name in the Dask graph to make things a bit
clearer.