Tweak apply_along_axis's pre-NumPy 1.13.0 error (#2703)
* Sum non-trivial dimensions as integers

NumPy will upcast the bool array that we are summing to integers as it
is a lower precision type. However there is no harm in forcing a higher
precision type to start with. Hence we force `sum` to convert the `bool`
array to `int` for summation.

* Tweak apply_along_axis's error message

The previous warning message implied that the returned array needed one
non-trivial dimension. However it could actually be a scalar. So instead
point out that more than one non-trivial dimension is an issue.