Add apply_over_axes (#2702)
* Add apply_over_axes to Dask Array

Implements a Dask Array equivalent of `apply_over_axes`. Does this by
simply calling `apply_along_axis` for each axis that `apply_over_axes`
was provided. Adds in a singleton dimension afterwards if the axis was
reduced. However if the function does not preserve the dimensionality or
reduce it exactly by one, `apply_over_axes` raises a `ValueError` like
its NumPy equivalent.

* Export apply_over_axes to Dask Array's public API

* Test Dask Array's apply_over_axes

Provide some tests for the Dask Array implementation of
`apply_over_axes`, which compares it to the NumPy implementation. Try a
range of different functions and axes they are applied along. Also
explores functions that return something with some dimensionality to
them.

* Document Dask Array's apply_over_axes