Drop s3fs requirement
Seems like `s3fs` is or can be optional as one can certainly run a
`distributed` cluster without using S3. Looking at `distributed`, it
appears to have dropped any requirement on `s3fs` at all. Also any
import in `dask` of `s3fs` is guarded against in case `s3fs` is not
present. As there are fallbacks that are commonly used when S3 is not
used, it seems reasonable to drop `s3fs` as a required dependency of the
`distributed` backend. Though maybe `s3fs` should live under a different
extra install option.