Add atleast_nd functions (#2760)
* Add atleast_1d

Implements the equivalent of NumPy's `atleast_1d` for Dask Arrays.

* Support 0-D arrays in atleast_2d

Previously the 0-D array case of `atleast_2d` was not handled. This
implements support for the 0-D case.

* Support 0-D arrays in atleast_3d

Previously the 0-D array case of `atleast_3d` was not handled. This
implements support for the 0-D case.

* Reuse NumPy docstrings for atleast_nd

The `atleast_2d` and `atleast_3d` functions were added some time ago,
but do not have docstrings. So just reuse the docstrings from NumPy per
usual. Also do the same for `atleast_1d`.

* Export atleast_nd functions to public API

The `atleast_2d` and `atleast_3d` functions were not part of the public
API previously. So this adds them to the public API to mirror NumPy.
Also adds the newly minted `atleast_1d` as well.

* Test atleast_nd functions

Simply compares the results of the Dask Array implementations to those
of NumPy's.

* Document atleast_nd

Includes documentation for `atleast_2d` and `atleast_3d`, which have
been around for a while. Also adds documentation for the newly minted
`atleast_1d`.

* Add changelog entry for atleast_nd functions

* Store atleast_nd expanded arrays before returning

* Handle multiple arrays in atleast_nd functions

* Test multiple arguments to atleast_nd

* Convert arguments to arrays in atleast_nd

Make sure we are dealing with Dask Arrays in all atleast_nd functions.