Follow-up for atleast_nd (#2765)
* Use `asanyarray` in `atleast_nd` functions

* Break out no argument tests of atleast_nd

Was including many combinations of no argument tests with other things
like array shapes that were not relevant. This simply runs the no
argument tests on the functions sans other parameters.

* Break out the one argument `atleast_nd` tests

This simplifies the test suite a bit by having the one argument case
dealt separately with than the two argument case.

* Test atleast_nd functions using two arrays

Make sure the two arrays can have different dimensionality to ensure
that both are being taken care of correctly regardless of whether their
dimensionality is the same or not.

* Skip some two arguments tests of atleast_nd

These cases are already tested with the arguments in reversed order. So
there is no need to test them again. Hence they are selectively skipped.

* Avoid using pytest's skip

Instead of skipping parameter combinations with `pytest.skip`, just
construct the combinations we care about using an `itertools` trick.