Add allclose (#2771)
* Add allclose

Implements the equivalent of NumPy's `allclose` for Dask Arrays. Does
this by simply using the existing `isclose` operation for comparing each
element. Then calls `all` on the result of `isclose` to collapse it down
to a single `bool`.

* Expose allclose to Dask Array's public API

* Tests `allclose`

Provides a very simple test of Dask Array's `allclose`, which compares
the result against NumPy's `allclose`. Basically borrows and tweaks a
test used for Dask Array's `isclose`.

* Adds Dask Array's allclose to the API docs

* Reference ``allclose`` in the changelog