ResourceProfiler plot works with single point
Previously if there was only a single point the (start, end) plot bounds
would be set to identical values, leading to js errors when trying to
view the plot. This fixes this issue and adds a test for it.

Note that this is only an issue for the ResourceProfiler - all other
plots already checked for this.