Implement Dask Array's unique to be lazy
Previously Dask Array's `unique` would compute the result when called.
Given Dask's typically lazy nature, this was a bit surprising for users.
However for a while the needed features for a lazy `unique`
implementation (e.g. unknown shapes) simply did not exist. Now that it
looks like the needed features are available, rewrite `unique` to be
lazy to more closely match the behavior of other functions in Dask.