Run `compute` on `unique` in `from_bcolz`
Appears that `from_bcolz` was using Dask Array's `unique` and expected
it to not be lazy. Making `unique` lazy broke it apparently. However as
the non-lazy version simply returned a NumPy array with the result, it
seems pretty reasonable to just have `from_bcolz` call `compute` on the
lazy `unique` result to get the expected NumPy result, which is what is
done here. The computational behavior should be the same as before.