Implement Dask Array's unique to be lazy (#2775)
* Implement Dask Array's unique to be lazy

Previously Dask Array's `unique` would compute the result when called.
Given Dask's typically lazy nature, this was a bit surprising for users.
However for a while the needed features for a lazy `unique`
implementation (e.g. unknown shapes) simply did not exist. Now that it
looks like the needed features are available, rewrite `unique` to be
lazy to more closely match the behavior of other functions in Dask.

* Test that unique is now lazy

* Update changelog to note unique is lazy

* Run `compute` on `unique` in `from_bcolz`

Appears that `from_bcolz` was using Dask Array's `unique` and expected
it to not be lazy. Making `unique` lazy broke it apparently. However as
the non-lazy version simply returned a NumPy array with the result, it
seems pretty reasonable to just have `from_bcolz` call `compute` on the
lazy `unique` result to get the expected NumPy result, which is what is
done here. The computational behavior should be the same as before.

* Construct `unique`'s reduction graph directly

Instead of making use of `delayed` to construct `unique`'s reduction
graph, simply hand roll our own reduction graph directly. This should
avoid unneeded calls to `optimize` and still work well for the end user.