Avoid double-scheduling in work stealing (#1489)
* Avoid double-scheduling in work stealing

This avoids double-scheduling tasks when work stealing by first checking
with the victim worker to verify that it has not yet started execution.
If it has started then we reverse the steal operation, if it has not
started then we cleanly start the new task only after the victim worker
has removed the task.

This passes most tests (those tests within stealing.py that if fails are
reasonable failures) but typically this sort of change triggers some
knock-on effects.

* fix tests

* avoid stealing confirmation if task state has changed

* avoid validation if stealing in flight

* update tests

* stealing story fits new logs

* Comment out occupancy/processing validation

I would like to get the new work stealing into master soon.  It does
sometimes result in incorrect occupancy calculations.  I should fix this
before the next release, but don't want to hold up near-term progress.

* clean up stealing

* handle in_flight

* Avoid worker death from stealing if task no longer present

* relax work stealing task

* Defend against KeyErrors in stealing