Support other arguments to `unique`
Implements support for `unique`'s arguments `return_index`,
`return_inverse`, and `return_counts`. All of which are optionally
supported and lazily computed with the unique values. The results are
returned as a `tuple` if any of the other arguments are set to `True`.
Otherwise behavior is identical if none of the arguments are set to
`True`. Matches with NumPy's `unique`'s behavior.

Also explicitly use `int64` for `unique`'s extra returns.  This is
needed on Windows, in particular, as Python uses `int32` by default.
However `unique` returns `int64` for all indices and counts on both Unix
and Windows platforms. So forcing `int64` will ensure Windows works
correctly and shouldn't change anything for Unix.