Vectorize `_unique_internal`'s `for`-loop
Makes use of NumPy's `where` to eliminate a `for`-loop over each unique
value found. Instead takes the original arrays and fills in values such
that the reduction applied should ignore them entirely (e.g. `0` for
`sum` and largest possible value for `min`). As there should always be
one value in the original array matching the unique value checked
(otherwise it wouldn't be a unique value), there should be a defined
result for `sum` and `min` that is not the fill value in both cases.

This strategy could be approached with `MaskedArray`s. However this was
tried previously and found to be notably (though not massively) slower
than simply using the fill values.