Error nicely on parse dates failure in read_csv
In pandas, if a column specified in `parse_dates` has an invalid value,
the column is silently converted to `object` dtype. This can cause
issues if some columns infer as datetimes and others as object dtype, as
converting with `astype` is prone to failure.

To remedy this, we add a nice error message if any partition expecting a
datetime dtype finds an object dtype instead.