Close resource profiler process on __exit__ (#2871)
Previously ``ResourceProfiler`` would start a tracker process on
`__init__`, which would then persist until `.close()` was called (or the
profiler was garbage collected). This led to processes unintentionally
persisting longer than intended by users.

Now we close the tracker process on context exit, and restart it if it's
closed on enter. When registered globally (not in a context manager) the
process persists until `close` is explictly called, as it was before.
All user code should continue to work the same, and the cost of
restarting the process should be minimal compared to the cost of the
computations being profiled.