Rewrite `matmul` to support N-D arrays
Previously `matmul` was restricted to 1-D and 2-D arrays. This rewrites
`matmul` to support N-D arrays with N > 1. Support for N > 2 is
accomplished by using `atop` with NumPy's `matmul`. Some adjustments to
the arrays are made to handle the 1-D array case correctly and handle
broadcasting in the same way that NumPy's `matmul` would.