Vendor NumPy 1.13's _Recurser in numpy_compat
Unfortunately we need to vendor `numpy.core.shape_base._Recurser` as it
appears to get dropped midway through NumPy 1.13 and is replaced with a
new implementation, which will require some fiddling to get working with
Dask Arrays. In comparison, the `numpy.core.shape_base._Recurser`
basically just works with any so is pretty easy to get working with Dask
Arrays as well.