Add block (#2650)
* Vendor NumPy 1.13's _Recurser in numpy_compat

Unfortunately we need to vendor `numpy.core.shape_base._Recurser` as it
appears to get dropped midway through NumPy 1.13 and is replaced with a
new implementation, which will require some fiddling to get working with
Dask Arrays. In comparison, the `numpy.core.shape_base._Recurser`
basically just works with any so is pretty easy to get working with Dask
Arrays as well.

* Add block

Provide a Dask Array implementation of NumPy's `block` function. This is
basically verbatim the code from NumPy with some tweak to use Dask Array's
equivalents of NumPy operations like `concatenate`.

* Export block to Dask Array's public API

* Test Dask Array's block

Inspired by NumPy's tests of `block` except changed to compare Dask
Array's implementation of `block` to NumPy's implementation. Some of the
error tests are basically identical to those in NumPy.

* Document block in Dask Array's API docs

* Note `block` in the changelog