Find subindex before acquiring lock
Within `insert_to_ooc`'s closure `store`, determine the `subindex` based
on the `region` and `index` before acquiring the lock. This can be done
independently of writing the data. So allow it to proceed in parallel
and be done before actually writing out the data. Should cutdown the
time the lock is held.