Add the `retrieve_from_ooc` function
Designs the `retrieve_from_ooc` to take the Dask graph created by
`insert_to_ooc` and create another Dask graph that loads the chunks
stored in `out` and returns them. This will allow Dask graphs to be
created for use as Dask Arrays that lazily load results written to
storage for later computation.