Changes `store` for `compute=True, keep=True`
Previously when `store` was provided `keep=True`, it did nothing when
`compute=True` instead just returning the Dask Array and associated
graph unchanged. However this goes against the grain of what
`compute=True` means. So this changes it such that `compute=True`
submits all of the storage jobs and reuses the `Future`s in creation of
the Dask Array returned. That way data can only be loaded if the results
have already been written. Though this provides the nice option of using
this result in downstream computation and/or further inspection before
proceeding.