Only create `Delayed` objects if `persist`ing
Instead of creating `Delayed` objects for all the cases that do not use
them (and merely unpack their contents), only create them when
`keep=True, compute=True` (i.e. when `persist`ing the storing step).
This cuts down on some overhead for the other cases and simplifies the
code a bit.