Optionally return stored data after storing (#2980)
* Test that `store` returns `None` normally

* Add a `keep` argument to `insert_to_ooc`

Provides a new argument for `insert_to_ooc` called `keep`, which
determines whether `out` is returned or not. The purpose being to chain
`out` into a later loading step for reading the data already written to
disk. By default `keep=False`, which preserves the existing behavior or
returning `None`. However if `keep=True`, then `out` is returned
instead.

* Add the `retrieve_from_ooc` function

Designs the `retrieve_from_ooc` to take the Dask graph created by
`insert_to_ooc` and create another Dask graph that loads the chunks
stored in `out` and returns them. This will allow Dask graphs to be
created for use as Dask Arrays that lazily load results written to
storage for later computation.

* Add optional `keep` argument to `store`

Provides the `store` function an optional `keep` argument. By default
`keep=False`, which provides the existing behavior as is. However if
`keep=True`, then a Dask Array is returned that with store the results
to `out` and reload them for later computation. For now ignores the
`compute` argument.

* Changes `store` for `compute=True, keep=True`

Previously when `store` was provided `keep=True`, it did nothing when
`compute=True` instead just returning the Dask Array and associated
graph unchanged. However this goes against the grain of what
`compute=True` means. So this changes it such that `compute=True`
submits all of the storage jobs and reuses the `Future`s in creation of
the Dask Array returned. That way data can only be loaded if the results
have already been written. Though this provides the nice option of using
this result in downstream computation and/or further inspection before
proceeding.

* Optimize `sources` before storing chunks

Puts all of the `sources` into one big Dask graph and then optimizes
them as Dask Arrays by call `__dask_optimize__` on all of the Array
sources with keys provided. Thus any optimizations that might be shared
between `sources` are performed only once for all of them. This graph is
then reused to reconstruct each `src` from the optimized graph before
storing them to each `tgt`.

As a bonus, this should take care of optimization for the `keep=False,
compute=False` case that were not previously handled. Also should ensure
that optimizations handled by `compute_as_if_collection` still get
handled.

* Combine all store graphs into one large graph

Instead of trying to handle each store graph separately, combine them
into one large graph that is used by all `Delayed` objects.

* Only create `Delayed` objects if `persist`ing

Instead of creating `Delayed` objects for all the cases that do not use
them (and merely unpack their contents), only create them when
`keep=True, compute=True` (i.e. when `persist`ing the storing step).
This cuts down on some overhead for the other cases and simplifies the
code a bit.

* Merge `load_dsks` into one shared graph

Instead of merging each of `load_dsks`' graphs within a `for`-loop,
merge them altogether before entering the `for`-loop and reuse that
graph for all Dask Arrays constructed.

* Cleanup returning of the result

Make sure there is only one `return` statement to make it easier to see
that a `return` does occur.

* Test storing and keeping results after

Includes some tests to exercise `store`'s `keep=True` cases to make sure
this behaves as intended. Namely that a Dask Array is returned and that
storage only occurs if `compute=True` or if `compute=False` and a later
call to `compute` triggers this.

* Change `keep` to `return_stored`

Follow a suggestion to rename the argument to make it a little clearer.

* Import Delayed at the top-level

* Import `persist` at top-level

* Unwrap code that fits on one line

* Add a comment to explain `retrieve_from_ooc`

Namely explain how arguments are reused from the `store_chunk` calls to
construct the `load_chunk` calls. Also explain what is dropped and
replaced in these new calls.

* Inline `return` calls in `if` branches

Should make it a little clearer what is returned under which conditions.

* Shorten comment about `load_dsk` construction

* Use comprehension for `tuple` of `Array`s

Follow-up on a suggestion to make this code more compact.