Fixes a metadata issue with store's return_stored (#3064)
* Test two arrays of different sizes and types

Make sure that each array's metadata is handled correctly and uniquely.

* Fix a bug in `store` when `return_stored=True`

There was an issue of applying the same metadata from the last array to
all preceding arrays when `return_stored=True`. This fixes that issue by
adding a `for`-loop and selecting out each array from `sources`
appropriately to match the data loaded from disk.

* Use generator expression to create arrays

To alleviate some style concerns raised, switch to using a generator
expression in a `tuple` using `zip`. Add an `assert` afterwards to
ensure we got the number of results we expected (same as number of
sources).

* Note `store`'s `return_stored` in the changelog

We forgot to include a changelog entry for the original PR. This
corrects that error.

* Add `store`'s metadata bug fix in the changelog