Fuse store steps into load steps
Calls of `load_chunk` following calls of `store_chunk` end up reusing
many of the same values. Namely the location where they store/load data
from, the lock they use (if any) to access the data, the region in the
output where the data lives, and the piece of the data they are
concerned with. As such, it makes sense to fuse these steps together
into one step to encourage the scheduler to run corresponding
`store_chunk` and `load_chunk` calls on the same worker. The fusing of
`store_chunk` and `load_chunk` calls is handled in two subtly different
ways depending on whether the `store_chunk` were called immediately or
whether they were delayed. In both cases `inline` and `cull` are used to
consolidate the graph eliminating independent `store_chunk` keys (and
possibly tasks) without changing the rest of the graph.

In the first case when handling `store_chunk` calls that were already
`persist`ed, Dask should return either `Future`s that reference the jobs
submitted to `distributed` (and their soon to be available results) or
should contain actual values if running locally. In either case, we want
to inline these `Future`s or results into the `load_chunk` calls in the
graph.  This is done by calling `inline` and `cull` to remove the now
unneeded keys from the graph. Thus allowing these results to be passed
as part of the `load_chunk` calls.

In the second case when Dask Arrays are requested from `store` and the
`store_chunk` operations are still in the graph, we simply want to
ensure that the corresponding `load_chunk` and `store_chunk` calls are
treated as one task and submitted to one worker. To do this, we again
use `inline` and `cull` to place all the `store_chunk` calls within
`load_chunk` calls and drop all `store_chunk` keys from the graph. This
way when submitting these jobs, each `load_chunk` gets passed with its
corresponding `store_chunk` call to one worker to compute the result.