Optionally keep lock in `store_chunk`/`load_chunk`
As the `store_chunk` and `load_chunk` tasks are now always fused
together to run on the same worker, there are some cases that would
benefit from continuing to keep the lock acquired in `store_chunk` so
that `load_chunk` can proceed immediately after without issues. In
particular this would occur if `store_chunk` calls were delayed. Thus
this adds a new parameter called `keep_lock` or `kept_lock`, which tells
`store_chunk` to not release the lock or `load_chunk` to skip acquiring
the lock. By default this parameter is set to `False`, which preserves
the existing locking behavior where `store_chunk` always releases the
lock and `load_chunk` always acquires it (assuming `load_chunk` is run).