Keep lock when `store_chunk`/`load_chunk` both run
When running `store_chunk` and `load_chunk` both as one task, make sure
to keep the lock acquired at the end of `store_chunk` and simply release
the lock at the end of `load_chunk`. This way the task won't be blocked
part way through due to contention over the lock. Instead the task
acquires the lock, does all of its work and releases the lock once it is
complete.