Rename `load_chunk` to `load_store_chunk`
As we are now fusing `store_chunk` and `load_chunk` operations into one
operation, it makes more sense to rename `load_chunk` to
`load_store_chunk`. This is a bit more descriptive of its behavior. Also
it should make more sense to users when inspecting the graph. As one
will not see the `store_chunk` step, seeing `load_store_chunk` will make
it more apparent that the store and load operations are being fused into
one step that is being run as a single task. Also this matches with the
existing `load_store_` that we prefix each key of the input data with.