Move `optimization` import out of `store`
Previously we were encountering `ImportError`s when trying to `import`
from `optimization` at the module level. So we moved the `import` calls
into the `store` function. However we are no longer seeing these
`ImportError`s when importing `optimization` at the global level. So go
ahead and move the import up.