Merge `store`'s target dasks
Before merging the target dasks into the store dasks, merge the target
dasks together. Typically this will be an empty `SharedDict`. However in
the cases where it might not be, it would be good to fuse all the target
dasks together so that optimizations can be performed on them as well
before combining them with the store dasks.