Cull unneeded keys from targets' dask
Using the keys collected for the targets' dask, cull anything that does
not directly rely on those keys. This will simplify the dask used for
the store step and make later optimization steps easier. As the typical
does not use `Delayed` objects for targets, this should be a no-op. Only
for the `Delayed` case will this clean up the graph a little bit much as
is already done with the sources' dask.