Allow the CI dev build to fail
As dev libraries tend to breaking things rather frequently, simply mark
this build as an allowed failure. This doesn't mean that we don't care
about fixing issues with dev libraries. Simply that we don't expect
every change to take that into account. Leaving it in the matrix for
those interested in fixing it to have an easier time doing so.