More tests of store's returned result (#3152)
* Clear delayed target before retesting store

* Move region tests with no result returned

Make sure all the region tests with no result returned are together.
This include with a single region and multiple regions.

* Wrap long line in tests

This line isn't over the limit, but it will soon be. So wrap it
proactively.

* Test immediate/delayed store and return of result

Was not testing whether `store` with result returned worked correctly
when the result was persisted to disk previously, this fixes that issue
by adding tests for this case.