Rewrite `to_hdf5` to be multiprocess friendly
Uses the `HDF5Dataset` wrapper class to provide an object that can be
used easily with `store` and adds in a `SerializableLock` to make
multiprocess writing possible. This still comes with the downside of
opening and closing files frequently and lots of locking with HDF5.
However it's much better than nothing for multiprocess usage, which is
the current situation.