Fuse region/chunk slices when creating store Dask
Instead of leaving the user-provided `region` and chunk slice (i.e.
`index`) to be fused in the `store_chunk` and `load_chunk` functions, go
ahead and fuse them when constructing the Dask graph and leave `region`
as `None`. This should cutdown the work done in `store_chunk` and
`load_chunk`. Also should allow for the `region` argument in both of
these functions to be subsequently dropped.