Drop region arg from store/load chunk functions
Since `fuse_slice` is now run as part of Dask construction and this
removes the need to pass the `region` on for handling in the store/load
chunk functions, simply drop the `region` arguments and related logic
from `store_chunk` and `load_chunk`.  Makes them a bit easier to follow
and maintain going forward.