Optimize `store`'s `targets` through `Delayed`
Use the standard way of optimizing Dask Delayed objects to optimize the
targets in advance if needed. Thus if `cull`ing or any other
optimization strategy becomes common place for Delayed objects, it will
be handled through the standard mechanisms instead of us doing it in
some custom way.

As an added bonus, this avoids collect empty dictionaries in the event a
target does not have a Dask (e.g. not Delayed).  Previously `store` had
collected empty dictionaries in this case. So this should be an
improvement when constructing and working with the `SharedDict` object
for `targets` in the common case (i.e. no Delayed targets).

Also provides a check to guarantee we are actually getting Dask Delayed
objects and are not supporting any other Dask collections. If we don't
get a Dask Delayed object, we assume this is some Array-like object.