Merge store Dasks together first
Before merging store Dasks with source and target Dasks, go ahead and
merge the store Dasks together alone. Simplifies the merging process a
bit and avoids directly relying on a dependency. Also makes sure that
there is a `SharedDict` for each group of related dictionaries that get
added to final store Dask. Namely the `sources`, `targets`, and now the
`store_chunk`s steps.