Add optional postcompute Dask to retrieve_from_ooc
Provides an option to give `retrieve_from_ooc` a computed Dask as well
to fill in values in the `load_chunk`s part of the Dask. This can be
useful when inlining results from the `persist`ed `store_chunk`s step.
This argument is totally optional and will be used to inline computed
results of `Future`s as the case may be. If this argument is not
provided, then the keys are simply inserted into the new Dask as was the
case before.