Check how many times `store` acquires a `Lock`
Verifies how many times a `Lock` instance is acquired in `store` when
using `return_stored`. Namely if `compute=True`, the lock will be
acquired two times for each chunk as storing and loading are separate
steps to allow for storing of chunks to occur asynchronously. However if
`compute=False`, the lock will be acquired only one time for each chunk
as store and load are fused into a single step that acquires the lock
for both parts. This acts as a check to ensure that the `compute=False`
case of `return_stored` does in fact fuse storing and loading into a
single step that holds the lock for both parts.